Fix Time to Integer conversion in Zipkin output
This commit fixes the conversion of Ruby Time objects to integer numbers of
microseconds since the UNIX epoch. Previously, the conversion used the `to_i`
method which truncates the timestamp to the nearest second. In order to
get sub-second precision, the `to_f` method must be used.

+++ profile-parser.rb
@@ -536,8 +536,13 @@ module PuppetProfiler
         result["parentId"] = Digest::SHA2.hexdigest(parent.last)[0..15]
       end
 
+      # Zipkin reports durations in microseconds and timestamps in microseconds
+      # since the UNIX epoch.
+      #
+      # NOTE: Time#to_i truncates to the nearest second. Using to_f is required
+      # for sub-second precision.
       unless span.start_time.nil?
-        result["timestamp"] = span.start_time.to_i * 10**6
+        result["timestamp"] = Integer(span.start_time.to_f * 10**6)
       end
       result["duration"] = Integer(span.time * 10**6)
